+++ requirements.txt
@@ -1,1 +1,5 @@
+httpx
+matplotlib
+numpy
 requests
+tqdm
